perf: form-bundle rollup pipeline (FORMS-24968) — collapse serial ES module hops
Port the HDFC-validated bundle optimization to the boilerplate so all downstream
EDS Forms projects inherit the performance gains:
- 12 JS fetches → 1 bundle + 2 shims (−11 requests)
- 5 serial import hops → 2 hops (parallel preload)
- Worker startup: −31% (afb-runtime.min.js instead of full source)

New files: rollup/form.rollup.config.js, blocks/form/form.source.js,
blocks/form/rules/index.source.js, scripts/swap-shims.js
Modified: form.js and rules/index.js → 1-line shims; head.html adds 5 modulepreloads;
package.json adds build/build:dev/build:form-bundle scripts; pre-commit shim guard added
