+++ head.html
@@ -1,4 +1,9 @@
 <meta name="viewport" content="width=device-width, initial-scale=1"/>
+<link rel="modulepreload" href="/blocks/form/form.js">
+<link rel="modulepreload" href="/blocks/form/form-bundle.min.js">
+<link rel="modulepreload" href="/blocks/form/constant.js">
+<link rel="modulepreload" href="/blocks/form/rules/model/afb-runtime.min.js">
+<link rel="modulepreload" href="/blocks/form/rules/model/afb-formatters.min.js">
 <script src="/scripts/aem.js" type="module"></script>
 <script src="/scripts/scripts.js" type="module"></script>
 <link rel="stylesheet" href="/styles/styles.css"/>
