+++ package.json
@@ -21,6 +21,9 @@
     "test:unit": "npm run build:json --silent && mocha",
     "test:e2e": "playwright test",
     "debug:e2e": "playwright test --project='chromium' --debug",
+    "build": "rollup -c rollup/form.rollup.config.js && node scripts/swap-shims.js --mode prod",
+    "build:dev": "MODE=dev rollup -c rollup/form.rollup.config.js && node scripts/swap-shims.js --mode dev",
+    "build:form-bundle": "rollup -c rollup/form.rollup.config.js",
     "update:formula": "rollup -c rollup/jsonformula.rollup.config.js",
     "update:core": "rollup -c rollup/af-core.rollup.config.js",
     "update:formatters": "rollup -c rollup/af-formatters.rollup.config.js",
